+++ src/routes/pages/dashboard.tsx
@@ -1,11 +1,11 @@
 import React, { useState }from "react"
 import { RouteComponentProps, withRouter } from "react-router-dom";
-import Playlist from '../Playlist'
+import Playlist from '../../components/Playlist'
 import Grid from '@material-ui/core/Grid';
 import { makeStyles } from '@material-ui/core/styles';
 import Button from '@material-ui/core/Button';
 import { logout } from '../../Login'
-import MasterPlaylist from "../MasterPlaylist";
+import MasterPlaylist from "../../components/MasterPlaylist";
 import { Playlist as PlaylistObj } from "../../types";
 
 export const useStyles = makeStyles((theme) => ({
